feat: реализована привязка вакансий к HR-владельцу
