+++ alembic/versions/12b998a5266f_create_initial_database_schema.py
@@ -1,0 +1,130 @@
+"""create initial database schema
+
+Revision ID: 12b998a5266f
+Revises: 
+Create Date: 2025-08-30 12:49:09.002380
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+from app.core.config import settings
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '12b998a5266f'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('company_name', sa.String(), nullable=True),
+    sa.Column('hashed_password', sa.String(), nullable=False),
+    sa.Column('role', sa.Enum('CANDIDATE', 'HR', name='role'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+    op.create_table('vacancies',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('job_title', sa.String(), nullable=False),
+    sa.Column('company_name', sa.String(), nullable=True),
+    sa.Column('location', sa.String(), nullable=True),
+    sa.Column('salary_range', sa.String(), nullable=True),
+    sa.Column('key_responsibilities', sa.Text(), nullable=True),
+    sa.Column('tech_stack', sa.Text(), nullable=True),
+    sa.Column('required_experience', sa.Text(), nullable=True),
+    sa.Column('hard_skills', sa.Text(), nullable=True),
+    sa.Column('soft_skills', sa.Text(), nullable=True),
+    sa.Column('education', sa.String(), nullable=True),
+    sa.Column('what_we_offer', sa.Text(), nullable=True),
+    sa.Column('owner_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_vacancies_id'), 'vacancies', ['id'], unique=False)
+    op.create_index(op.f('ix_vacancies_job_title'), 'vacancies', ['job_title'], unique=False)
+    op.create_table('applications',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('status', sa.Enum('SCREENING', 'INTERVIEW_PENDING', 'COMPLETED', 'REJECTED', name='applicationstatus'), nullable=False),
+    sa.Column('resume_md', sa.Text(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('vacancy_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.ForeignKeyConstraint(['vacancy_id'], ['vacancies.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_applications_id'), 'applications', ['id'], unique=False)
+    op.create_table('evaluation_criteria',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('criterion', sa.String(), nullable=False),
+    sa.Column('weight', sa.Integer(), nullable=False),
+    sa.Column('vacancy_id', sa.Integer(), nullable=False),
+    sa.Column('owner_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
+    sa.ForeignKeyConstraint(['vacancy_id'], ['vacancies.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_evaluation_criteria_id'), 'evaluation_criteria', ['id'], unique=False)
+    op.create_table('application_screening_results',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('result_json', sa.JSON(), nullable=False),
+    sa.Column('application_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['application_id'], ['applications.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('application_id')
+    )
+    op.create_index(op.f('ix_application_screening_results_id'), 'application_screening_results', ['id'], unique=False)
+    op.create_table('interview_sessions',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('status', sa.Enum('SCHEDULED', 'IN_PROGRESS', 'COMPLETED', 'ERROR', name='interviewstatus'), nullable=False),
+    sa.Column('started_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
+    sa.Column('ended_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('application_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['application_id'], ['applications.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('application_id')
+    )
+    op.create_index(op.f('ix_interview_sessions_id'), 'interview_sessions', ['id'], unique=False)
+    op.create_table('interview_transcripts',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('role', sa.Enum('AVATAR', 'CANDIDATE', name='transcriptrole'), nullable=False),
+    sa.Column('message', sa.Text(), nullable=False),
+    sa.Column('timestamp', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
+    sa.Column('session_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['session_id'], ['interview_sessions.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_interview_transcripts_id'), 'interview_transcripts', ['id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_interview_transcripts_id'), table_name='interview_transcripts')
+    op.drop_table('interview_transcripts')
+    op.drop_index(op.f('ix_interview_sessions_id'), table_name='interview_sessions')
+    op.drop_table('interview_sessions')
+    op.drop_index(op.f('ix_application_screening_results_id'), table_name='application_screening_results')
+    op.drop_table('application_screening_results')
+    op.drop_index(op.f('ix_evaluation_criteria_id'), table_name='evaluation_criteria')
+    op.drop_table('evaluation_criteria')
+    op.drop_index(op.f('ix_applications_id'), table_name='applications')
+    op.drop_table('applications')
+    op.drop_index(op.f('ix_vacancies_job_title'), table_name='vacancies')
+    op.drop_index(op.f('ix_vacancies_id'), table_name='vacancies')
+    op.drop_table('vacancies')
+    op.drop_index(op.f('ix_users_id'), table_name='users')
+    op.drop_index(op.f('ix_users_email'), table_name='users')
+    op.drop_table('users')
+    # ### end Alembic commands ###
