--- alembic/versions/37cfe61559b1_create_interview_session_and_transcript_.py
@@ -1,56 +1,0 @@
-"""create interview session and transcript tables
-
-Revision ID: 37cfe61559b1
-Revises: 4b1b6a60d30d
-Create Date: 2025-08-29 20:50:52.608584
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-from app.core.config import settings
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '37cfe61559b1'
-down_revision: Union[str, Sequence[str], None] = '4b1b6a60d30d'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('interview_sessions',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('status', sa.Enum('SCHEDULED', 'IN_PROGRESS', 'COMPLETED', 'ERROR', name='interviewstatus'), nullable=False),
-    sa.Column('started_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-    sa.Column('ended_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('application_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['application_id'], ['applications.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('application_id')
-    )
-    op.create_index(op.f('ix_interview_sessions_id'), 'interview_sessions', ['id'], unique=False)
-    op.create_table('interview_transcripts',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('role', sa.Enum('AVATAR', 'CANDIDATE', name='transcriptrole'), nullable=False),
-    sa.Column('message', sa.Text(), nullable=False),
-    sa.Column('timestamp', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-    sa.Column('session_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['session_id'], ['interview_sessions.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_interview_transcripts_id'), 'interview_transcripts', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_interview_transcripts_id'), table_name='interview_transcripts')
-    op.drop_table('interview_transcripts')
-    op.drop_index(op.f('ix_interview_sessions_id'), table_name='interview_sessions')
-    op.drop_table('interview_sessions')
-    # ### end Alembic commands ###
