--- alembic/versions/4b1b6a60d30d_create_application_table.py
@@ -1,45 +1,0 @@
-"""create application table
-
-Revision ID: 4b1b6a60d30d
-Revises: c8aec7d64b14
-Create Date: 2025-08-29 12:46:39.884119
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-from app.core.config import settings
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4b1b6a60d30d'
-down_revision: Union[str, Sequence[str], None] = 'c8aec7d64b14'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('applications',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('status', sa.Enum('SCREENING', 'INTERVIEW_PENDING', 'COMPLETED', 'REJECTED', name='applicationstatus'), nullable=False),
-    sa.Column('resume_md', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('vacancy_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.ForeignKeyConstraint(['vacancy_id'], ['vacancies.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_applications_id'), 'applications', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_applications_id'), table_name='applications')
-    op.drop_table('applications')
-    # ### end Alembic commands ###
