--- alembic/versions/5d1ad221285c_create_user_table.py
@@ -1,45 +1,0 @@
-"""create user table
-
-Revision ID: 5d1ad221285c
-Revises: 
-Create Date: 2025-08-28 19:46:37.324988
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-from app.core.config import settings
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5d1ad221285c'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('company_name', sa.String(), nullable=True),
-    sa.Column('hashed_password', sa.String(), nullable=False),
-    sa.Column('role', sa.Enum('CANDIDATE', 'HR', name='role'), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_table('users')
-    # ### end Alembic commands ###
