--- alembic/versions/c8aec7d64b14_create_vacancy_and_evaluation_criterion_.py
@@ -1,62 +1,0 @@
-"""create vacancy and evaluation_criterion tables
-
-Revision ID: c8aec7d64b14
-Revises: 5d1ad221285c
-Create Date: 2025-08-28 21:39:38.161634
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-from app.core.config import settings
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c8aec7d64b14'
-down_revision: Union[str, Sequence[str], None] = '5d1ad221285c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('vacancies',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('job_title', sa.String(), nullable=False),
-    sa.Column('company_name', sa.String(), nullable=True),
-    sa.Column('location', sa.String(), nullable=True),
-    sa.Column('salary_range', sa.String(), nullable=True),
-    sa.Column('key_responsibilities', sa.Text(), nullable=True),
-    sa.Column('tech_stack', sa.Text(), nullable=True),
-    sa.Column('required_experience', sa.Text(), nullable=True),
-    sa.Column('hard_skills', sa.Text(), nullable=True),
-    sa.Column('soft_skills', sa.Text(), nullable=True),
-    sa.Column('education', sa.String(), nullable=True),
-    sa.Column('what_we_offer', sa.Text(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_vacancies_id'), 'vacancies', ['id'], unique=False)
-    op.create_index(op.f('ix_vacancies_job_title'), 'vacancies', ['job_title'], unique=False)
-    op.create_table('evaluation_criteria',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('criterion', sa.String(), nullable=False),
-    sa.Column('weight', sa.Integer(), nullable=False),
-    sa.Column('vacancy_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['vacancy_id'], ['vacancies.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_evaluation_criteria_id'), 'evaluation_criteria', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_evaluation_criteria_id'), table_name='evaluation_criteria')
-    op.drop_table('evaluation_criteria')
-    op.drop_index(op.f('ix_vacancies_job_title'), table_name='vacancies')
-    op.drop_index(op.f('ix_vacancies_id'), table_name='vacancies')
-    op.drop_table('vacancies')
-    # ### end Alembic commands ###
