--- alembic/versions/d801398c0571_create_application_screening_result_.py
@@ -1,42 +1,0 @@
-"""create application screening result table
-
-Revision ID: d801398c0571
-Revises: 37cfe61559b1
-Create Date: 2025-08-29 22:04:56.131673
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-from app.core.config import settings
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd801398c0571'
-down_revision: Union[str, Sequence[str], None] = '37cfe61559b1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('application_screening_results',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('result_json', sa.JSON(), nullable=False),
-    sa.Column('application_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['application_id'], ['applications.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('application_id')
-    )
-    op.create_index(op.f('ix_application_screening_results_id'), 'application_screening_results', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_application_screening_results_id'), table_name='application_screening_results')
-    op.drop_table('application_screening_results')
-    # ### end Alembic commands ###
