+++ app/repositories/vacancy_repository.py
@@ -8,15 +8,12 @@ class VacancyRepository:
     def __init__(self, db: Session):
         self.db = db
 
-    def create_vacancy(self, vacancy_data: VacancyCreate) -> Vacancy:
+    def create_vacancy(self, vacancy_data: VacancyCreate, owner_id: int) -> Vacancy:
         vacancy_dict = vacancy_data.model_dump(exclude={'evaluation_criteria'})
-        db_vacancy = Vacancy(**vacancy_dict)
+        db_vacancy = Vacancy(**vacancy_dict, owner_id=owner_id)
 
         for criterion_data in vacancy_data.evaluation_criteria:
-            db_criterion = EvaluationCriterion(
-                criterion=criterion_data.criterion,
-                weight=criterion_data.weight
-            )
+            db_criterion = EvaluationCriterion(**criterion_data.model_dump())
             db_vacancy.evaluation_criteria.append(db_criterion)
 
         self.db.add(db_vacancy)
